+++ pckEDA/no_supervisado/__init__.py
@@ -2,5 +2,8 @@ from .base import NoSupervisado
 from .clustering import Clustering
 from .acp import ACP
 from .hac import HAC
+from .kmeans import KMeans
+from .tsne import TSNE
+from .umap import UMAP
 
-__all__ = ["NoSupervisado", "Clustering", "ACP", "HAC"]
+__all__ = ["NoSupervisado", "Clustering", "ACP", "HAC", "KMeans", "TSNE", "UMAP"]